Move production functions to test folder extending public API for separated testability
